add editor
