+++ src/pages/_app.tsx
@@ -1,6 +1,7 @@
 import { AppProps } from "next/app";
 import NextHead from "next/head";
 import { RecoilRoot } from "recoil";
+
 export default function App({ Component, pageProps }: AppProps) {
   return (
     <>
