+++ src/stores/story.ts
@@ -1,14 +1,12 @@
 import { atom } from "recoil";
 import { Story } from "~/modules/entity";
 
-interface State {
-  id: string;
-  data: Story;
-}
-
-const state = atom<State>({
-  key: "story",
+export const currentStoryIDState = atom<string>({
+  key: "currentStoryID",
   default: undefined,
 });
 
-export default state;
+export const currentStoryState = atom<Story>({
+  key: "currentStory",
+  default: undefined,
+});
